add class base with pulic method get_styles
