+++ pages/base.py
@@ -6,13 +6,63 @@ from selenium.webdriver.common.by import By, ByType
 from selenium.webdriver.remote.webdriver import WebDriver
 from selenium.webdriver.remote.webelement import WebElement
 
+from utilities.logger import Logger
+
 __all__ = ["BasePage", "BaseComponent", "LocatorType", "DictLocatorType"]
 
 LocatorType = Tuple[ByType, str]
 DictLocatorType = Dict[str, LocatorType]
 
 
-class BasePage:
+class Base:
+    """Utility class for CSS style-related operations."""
+
+    def __init__(self, driver: WebDriver):
+        """Initialize StylesGetter with a WebDriver instance.
+
+        Args:
+            driver: Selenium WebDriver instance.
+        """
+        self.driver = driver
+        self.logger = Logger.get_logger("StylesGetter")
+
+    def _get_computed_style(self, element: WebElement, property_name: str) -> str:
+        """Get computed CSS style value for an element.
+
+        Args:
+            element: WebElement to check
+            property_name: CSS property name (use camelCase format)
+
+        Returns:
+            str: Computed style value
+        """
+        self.logger.debug(f"Getting computed style: {property_name}")
+        script = "return window.getComputedStyle(arguments[0])[arguments[1]];"
+        return self.driver.execute_script(script, element, property_name)
+
+    def get_styles(self, element: WebElement, properties: dict) -> dict:
+        """Get multiple computed CSS styles for an element.
+
+        Args:
+            element: WebElement to check
+            properties: Dict where keys are property names (camelCase) to retrieve
+
+        Returns:
+            dict: Dictionary of property names and their values
+        """
+        if not element:
+            self.logger.debug("Cannot get styles: element is None")
+            return {}
+
+        self.logger.debug("Getting multiple styles")
+        styles = {}
+        for prop in properties:
+            styles[prop] = self._get_computed_style(element, prop)
+        self.logger.debug(f"Styles retrieved: {styles}")
+        return styles
+
+
+class BasePage(Base):
     """Base class for all page objects."""
 
     locators: DictLocatorType = {"header": (By.CSS_SELECTOR, "#app > ul")}
@@ -76,7 +126,7 @@ class BasePage:
         return self.driver.find_elements(*locator)
 
 
-class BaseComponent:
+class BaseComponent(Base):
     """Base class for all page components."""
 
     def __init__(self, driver: WebDriver, parent: WebElement) -> None:
