+++ utilities/__init__.py
@@ -1,6 +1,5 @@
 """Set Utilities package for test automation project."""
 
 from .logger import Logger, get_logger
-from .styles_getter import StylesGetter
 
-__all__ = ["Logger", "get_logger", "StylesGetter"]
+__all__ = ["Logger", "get_logger"]
